Updated how attendance is stored

+++ setPremData.sql
@@ -30,7 +30,7 @@ CREATE TABLE matches (
     homeXG DECIMAL(2, 1),
     awayGoals INT NOT NULL,
     awayXG DECIMAL(2, 1),
-    attendance INT,
+    attendance DECIMAL(6, 1),
     location VARCHAR(200),
     
     PRIMARY KEY(matchID),
